+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- `sector_id` and `muni_id` from `exploitation` table.
+
 ## [4.10.0] - 2026-05-18
 
 ### Added